separating host keys by type
