+++ modules/core/24-host/connection.go
@@ -1,0 +1,28 @@
+package host
+
+import "fmt"
+
+const KeyConnectionPrefix = "connections"
+
+// ICS03
+// The following paths are the keys to the store as defined in https://github.com/cosmos/ibc/blob/master/spec/core/ics-003-connection-semantics#store-paths
+
+// ClientConnectionsPath defines a reverse mapping from clients to a set of connections
+func ClientConnectionsPath(clientID string) string {
+	return FullClientPath(clientID, KeyConnectionPrefix)
+}
+
+// ClientConnectionsKey returns the store key for the connections of a given client
+func ClientConnectionsKey(clientID string) []byte {
+	return []byte(ClientConnectionsPath(clientID))
+}
+
+// ConnectionPath defines the path under which connection paths are stored
+func ConnectionPath(connectionID string) string {
+	return fmt.Sprintf("%s/%s", KeyConnectionPrefix, connectionID)
+}
+
+// ConnectionKey returns the store key for a particular connection
+func ConnectionKey(connectionID string) []byte {
+	return []byte(ConnectionPath(connectionID))
+}
